OUTPUT: including the values of all the dead players.
In the previous versions, the scores of the dead players were not included, invalidating the table-separated format.

I had to add a dictionary to keep track of all the dead players and their scores.

KNOWN BUG: the reputation of all the dead players keep decreasing after their death, because self.total_attempts increases